requirements.txt
psutil==5.9.8
numpy==1.26.4
python-dotenv==1.0.1
kafka-python==2.0.2
